+++ src/editors/SampledValues.ts
@@ -4,6 +4,7 @@ import '@material/mwc-fab';
 
 import './sampledvalues/subscriber-ied-list-smv.js';
 import './sampledvalues/smv-publisher-list.js';
+import './sampledvalues/smv-subscriber-list.js';
 import { RadioListItem } from '@material/mwc-list/mwc-radio-list-item.js';
 import { newViewEvent, View, ViewEvent } from './foundation.js';
 import { translate } from 'lit-translate';
@@ -61,7 +62,7 @@ export default class SampledValuesPlugin extends LitElement {
       <div class="container">
         ${view == View.PUBLISHER
           ? html`<smv-publisher-list class="row" .doc=${this.doc}></smv-publisher-list>`
-          : html`<smv-publisher-list class="row" .doc=${this.doc}></smv-publisher-list>`
+          : html`<smv-subscriber-list class="row" .doc=${this.doc}></smv-subscriber-list>`
         }
         <subscriber-ied-list-smv
           class="row"
