+++ src/editors/foundation.ts
@@ -5,6 +5,15 @@ export enum View {
   SUBSCRIBER
 }
 
+/**
+ * Enumeration stating the Subscribe status of a IED to a GOOSE or Sampled Value.
+ */
+ export enum SubscribeStatus {
+  Full,
+  Partial,
+  None,
+}
+
 export interface ViewDetail {
   view: View;
 }
@@ -21,10 +30,20 @@ export function newViewEvent(
   });
 }
 
-declare global {
-  interface ElementEventMap {
-    ['view']: ViewEvent;
-  }
+export interface IEDSelectDetail {
+  ied: Element | undefined;
+}
+export type IEDSelectEvent = CustomEvent<IEDSelectDetail>;
+export function newIEDSelectEvent(
+  ied: Element | undefined,
+  eventInitDict?: CustomEventInit<IEDSelectDetail>
+): IEDSelectEvent {
+  return new CustomEvent<IEDSelectDetail>('ied-select', {
+    bubbles: true,
+    composed: true,
+    ...eventInitDict,
+    detail: { ied, ...eventInitDict?.detail },
+  });
 }
 
 export function getOrderedIeds(doc: XMLDocument): Element[] {
@@ -33,4 +52,11 @@ export function getOrderedIeds(doc: XMLDocument): Element[] {
         compareNames(a, b)
       )
     : [];
+}
+
+declare global {
+  interface ElementEventMap {
+    ['view']: ViewEvent;
+    ['ied-select']: IEDSelectEvent;
+  }
 }
